pyproject.toml
[project]
name = "maui"
version = "0.1.0"
description = "Maui: Intelligent Task Assistant"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot[job-queue]>=20.7",
    "google-generativeai>=0.3.2",
    "peewee>=3.17.0",
    "pydantic>=2.5.3",
    "python-dotenv>=1.0.0",
    "pytz>=2023.3",
    "rich>=13.0.0",
    "textual>=0.47.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
    "ruff>=0.14.8",
]
